fix: proper handling for .bat/.cmd on windows
